Use verify_access to do authorization on all protected handlers
